Refactor: Extract base class URL
